pyproject.toml
[tool.poetry]
name = "solana-as-api"
version = "0.1.0"
description = "reference implementation of using solana addresses and pkrivate key signing for api auth"
authors = ["tamgros <sorg.matthew@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
solana = "^0.29.2"
fastapi = "^0.95.1"
solders = "^0.14.4"
pynacl = "^1.5.0"
uvicorn = "^0.21.1"
pydantic = "^1.10.7"
python-jose = "^3.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:start"
